fix lint and moved Given to common steps
