+++ features/purpose/step_definitions/purpose-archive.ts
@@ -1,16 +1,10 @@
-import { Given, When } from "@cucumber/cucumber";
+import { When } from "@cucumber/cucumber";
 import { z } from "zod";
 import {
   assertContextSchema,
   getAuthorizationHeader,
-  getToken,
 } from "../../../utils/commons";
 import { apiClient } from "../../../api";
-import { TenantType } from "../../common-steps";
-import {
-  ESERVICE_DAILY_CALLS,
-  dataPreparationService,
-} from "../../../services/data-preparation.service";
 import { PurposeVersionState } from "../../../api/models";
 
 When(
@@ -31,24 +25,5 @@ When(
       versionId!,
       getAuthorizationHeader(this.token)
     );
-  }
-);
-
-Given(
-  "{string} ha già richiesto l'aggiornamento della stima di carico superando i limiti di quell'e-service",
-  async function (tenant: TenantType) {
-    assertContextSchema(this, {
-      purposeId: z.string(),
-    });
-
-    const token = getToken(this.tokens, tenant, "admin");
-
-    const { waitingForApprovalVersionId } =
-      await dataPreparationService.createNewPurposeVersion(
-        token,
-        this.purposeId,
-        { dailyCalls: ESERVICE_DAILY_CALLS.perConsumer + 1 }
-      );
-    this.waitingForApprovalVersionId = waitingForApprovalVersionId;
   }
 );
